% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{amat_to_bm}
\alias{amat_to_bm}
\title{Convert adjacency matrix into BoolModel object}
\usage{
amat_to_bm(amat, random = F)
}
\arguments{
\item{amat}{matrix. directed adjacency matrix.}

\item{random}{logical. Randomly assign to either act or inh rules, when the adjacency matrix only has values 0 and 1, but not -1.}
}
\description{
This function converts adjacency matrix to BoolModel object. Able to take in adjacency matrix with -1, which encodes for inhibitory interaction.
}

