% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{bin_to_real}
\alias{bin_to_real}
\title{Generate random real numbers from binary values}
\usage{
bin_to_real(x, param)
}
\arguments{
\item{x}{logical or 0/1 numeric matrix. Col-genes, row-samples.}

\item{param}{list of parameters given by param_bimodal().}
}
\description{
This function generates random real numbers from binary values, with supplied parameters. Returns a vector of real values.
}

