% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bm_to_amat}
\alias{bm_to_amat}
\title{Convert BoolModel into adjacency matrix}
\usage{
bm_to_amat(x, directed = T)
}
\arguments{
\item{x}{S4 BoolModel object.}

\item{directed}{logical. Whether to return directed or undirected adjacency matrix. Default to TRUE.}
}
\description{
This function converts a BoolModel object into an adjacency matrix.
}

