% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{bm_to_df}
\alias{bm_to_df}
\title{Convert BoolModel object into BoolNet readable data frame}
\usage{
bm_to_df(bmodel)
}
\arguments{
\item{bmodel}{BoolModel object.}
}
\description{
This method converts BoolModel object into a data frame, which is readable by BoolNet.
}

