% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{bon_bmodel}
\alias{bon_bmodel}
\title{HSC Boolean Model from Bonzanni et al.}
\format{A data frame with 11 rows and 2 columns. 

Rows: each row consists of 1 gene and its associated Boolean rule.
Column 1: target gene
Column 2: associated Boolean rule}
\usage{
data(bon_bmodel)
}
\description{
A Boolean model describing HSC in mice. It contains 11 genes. Its steady state is a cyclic loop of 32 states.
}

