% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{bon_moig_istate}
\alias{bon_moig_istate}
\title{Initial state from Moignard et al.}
\format{A data frame with 1 row and 20 columns.

Rows: each row consists of 1 set of Boolean state.
Columns: each column is for 1 gene/variable.}
\usage{
data(bon_moig_istate)
}
\description{
An intial state obtained from data in Moignard et al, determined by taking colMeans over unique rows, and rounding the means to 0-1.
Values for genes that are missing in Moignard et al, but are present in Bonzanni et al, are determined by taking values from the original initial state supplied in Bonzanni et al.
It contains a set of Boolean values for 20 genes.
}

