% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolmodel_class.R
\docType{class}
\name{BoolModel-class}
\alias{BoolModel}
\alias{BoolModel-class}
\title{An S4 class to represent a Boolean Model}
\description{
This class represents Boolean Model in a S4 BoolModel object.
}
\section{Fields}{

\describe{
\item{\code{target}}{character vector. It should contain gene symbols.}

\item{\code{targer_var}}{character vector. It should contain the internal representation of gene variables, in the form of "v[0-9]+s"}

\item{\code{rule_act}}{list of character vectors. Each element in the list should contain the activating gene variables for a particular target gene.}

\item{\code{rule_inh}}{list of character vectors. Each element in the list should contain the inhibiting gene variables for a particular target gene.}
}}

