% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/booltrainer.R
\docType{package}
\name{BoolTraineR}
\alias{BoolTraineR}
\alias{BoolTraineR-package}
\title{BoolTraineR: A package for studying asynchronous Boolean models}
\description{
This package contains tools for Boolean model manipulation, as well as the search for the best Boolean model.
}

