% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btr.R
\docType{package}
\name{BTR}
\alias{BTR}
\alias{BTR-package}
\title{BTR: A package for studying asynchronous Boolean models}
\description{
This package contains tools for Boolean model manipulation, as well as the search for the best Boolean model.
}

