% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculation.R
\name{calc_roc}
\alias{calc_roc}
\title{Calculate precision, recall, f-score, accuracy and specificity}
\usage{
calc_roc(x)
}
\arguments{
\item{x}{integer vector. Vector output by validate_adjmat().}
}
\description{
This function calculates the precision, recall, f-score, accuracy and specificity from the output of validate_adjmat().
}

