% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{check_bmodel}
\alias{check_bmodel}
\title{Check if the Boolean model violates constraints.}
\usage{
check_bmodel(bmodel, max_varperrule)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}
}
\description{
This function checks if the Boolean model violates contraints. Return logical value.
(1) Each gene rule should not have more terms than max_varperrule.
(2) The same term should not occur twice in the same rule.
}

