% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_bmodel}
\alias{compress_bmodel}
\title{Compress BoolModel}
\usage{
compress_bmodel(bmodel, encoding, max_varperrule)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{encoding}{named numerical vector returned by get_encodings().}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}
}
\description{
This function compresses S4 BoolModel object by representing variables using numbers, and also only the act rules and inh rules are kept. 
Return a list of 3 vectors, corresponding to act rules and inh rules.
}

