% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{decompress_bmodel}
\alias{decompress_bmodel}
\title{Decompress BoolModel}
\usage{
decompress_bmodel(x, encoding, format = "bmodel")
}
\arguments{
\item{x}{vector returned by compress_bmodel.}

\item{encoding}{named numerical vector returned by get_encodings().}

\item{format}{character. Specifies which format to return. Possible values: 'bmodel', 'df', 'amat', 'simp_df'. Default to 'bmodel'.}
}
\description{
This function decompresses the bmodel compressed by compress_bmodel(). 
Return a S4 BoolModel object.
}

