% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{decreate_boolmodel}
\alias{decreate_boolmodel}
\title{Decreate Boolean model}
\usage{
decreate_boolmodel(bmodel)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}
}
\description{
This function converts a S4 BoolModel object into a list of 4 lists. (for use by Rcpp in simulate_model())
}

