% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{emodel2}
\alias{emodel2}
\title{Example Boolean Model used in the vignette}
\format{Each Boolean model is a BoolModel object.}
\usage{
data(example_models)
}
\description{
A Boolean model used in the examples of the vignette.
}

