% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{equi_model}
\alias{equi_model}
\title{Check for equivalent models}
\usage{
equi_model(bmodel1, bmodel2, inter_bool, max_varperrule)
}
\arguments{
\item{bmodel1}{S4 BoolModel object.}

\item{bmodel2}{S4 BoolModel object.}

\item{inter_bool}{logical. Indicate whether to consider AND terms.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}
}
\description{
This function checks if the two models have the same rules. Return a logical value. Only TRUE if each rule for each gene is the same.
}

