% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{eval_bool}
\alias{eval_bool}
\title{Evaluating Boolean rules}
\usage{
eval_bool(bmodel, val, ind)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{val}{named logical vector. It should contain the values for all genes at that time point. Note that each value in the vector must be named by its corresponding gene name.}

\item{ind}{integer. It indicates the state of which gene should be computed.}
}
\description{
This function evaluates the Boolean rules (both act and inh) of one gene at a time. Return a logical value for that gene.
}

