% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{example_models}
\alias{example_models}
\title{Example Boolean Models used in the vignette}
\format{Each Boolean model is a BoolModel object.}
\usage{
data(example_models)
}
\description{
3 Boolean models used in the examples of the vignette.
}

