% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{filter_dflist}
\alias{filter_dflist}
\title{Filter columns of df in a list}
\usage{
filter_dflist(x, y, uniq_bool = T)
}
\arguments{
\item{x}{data frame. It should be a list element if used from lapply.}

\item{y}{character vector. It should contains gene names found in the colnames of the data frame.}

\item{uniq_bool}{logical. Whether to return unique rows only.}
}
\description{
This function filters columns of multiple df in a list, when compared using a vector. Use through lapply().
}

