% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rand_model.R
\name{gen_randata}
\alias{gen_randata}
\title{Generate sets of random data}
\usage{
gen_randata(n, steps, num_genes, max_varperrule, inter_bool)
}
\arguments{
\item{n}{integer. Number of sets of random data required.}

\item{steps}{integer. Specify the number of steps between the two Boolean models. If steps=0, give completely random starting model.}

\item{num_genes}{integer. Number of genes in the Boolean models.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}

\item{inter_bool}{logical. Indicate whether to consider AND terms.}
}
\description{
This function generates specified sets of random data, which include initial states, two Boolean models (starting and destination), and continuous+discrete data of the destination model.
}

