% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_model.R
\name{gen_singlerule}
\alias{gen_singlerule}
\title{Generate random act and inh rule for a single gene}
\usage{
gen_singlerule(x, np, tar_ind, and_bool, self_loop = F,
  rule_type = "random")
}
\arguments{
\item{x}{character vector. A vector of all single terms to be used.}

\item{np}{integer. Number of gene variables in a rule. NOT max_varperrule here.}

\item{tar_ind}{numerical. Indicate which gene is the rule for. Used in preventing self-loop.}

\item{and_bool}{logical. Indicates whether to include AND terms or not.}

\item{self_loop}{logical. Indicates whether to allow self_loop. Default to F.}

\item{rule_type}{character. Types of rules. Defaults to random.}
}
\description{
This function generates one random Boolean rule (both act and inh) per run. Return a list of two vectors.
Note that this method will not give empty rule, i.e. 0 term in both act and inh rules.
}

