% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compression.R
\name{get_encodings}
\alias{get_encodings}
\title{Get corresponding encodings for compression or decompression.}
\usage{
get_encodings(bmodel, inter_bool)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{inter_bool}{logical. Indicate whether to consider AND terms.}
}
\description{
This function gets all possible terms (single or double terms) and their corresponding encodings.
This function limits the number of possible variables in the model to 999.
}

