% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_modification.R
\name{grow_bmodel}
\alias{grow_bmodel}
\title{Add extra genes to a Boolean model}
\usage{
grow_bmodel(in_model, in_gene)
}
\arguments{
\item{in_model}{data frame with 2 columns, which are targets and factors}

\item{in_gene}{character vector. Genes to be added into the model.}
}
\description{
This function adds extra genes to a Boolean model. Input model must be in data frame format, output model will be BoolModel object.
}

