% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{initialise_data}
\alias{initialise_data}
\title{Initialise data}
\usage{
initialise_data(state, aslogic = F)
}
\arguments{
\item{state}{data frame. It should contain either 0/1 or F/T data of gene expression.}

\item{aslogic}{logical. It specifies whether to convert the input data into Boolean values. Default to FALSE.}
}
\description{
This function initialises data frame of Boolean state space. Returns initialised data frame.
}

