% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialisation.R
\name{initialise_raw_data}
\alias{initialise_raw_data}
\title{Initialise raw data}
\usage{
initialise_raw_data(x, data_type = "qpcr", uni_thre = 0.2, scale = T)
}
\arguments{
\item{x}{matrix. Numeric data of gene expression.}

\item{data_type}{character. Specify data types: qpcr, rnaseq.}

\item{uni_thre}{numerical. Speficy threshold for unimodality test. Default to 0.2.}

\item{scale}{logical. Whether to scale the data to a range of 0-1. Default to T.}
}
\description{
This function initialise raw gene expression values in a matrix. Return a list of two matrices: (1) continuous values and (2) binary values.
Note that kmeans clustering as binarisation only works well if the data has a bimodal distribution.
}

