% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{krum_bmodel}
\alias{krum_bmodel}
\title{Myeloid Boolean Model from Krumsiek et al.}
\format{A data frame with 11 rows and 2 columns. 

Rows: each row consists of 1 gene and its associated Boolean rule.
Column 1: target gene
Column 2: associated Boolean rule}
\usage{
data(krum_bmodel)
}
\description{
A Boolean model describing myeloid development in mice. It contains 11 genes. Its steady states are 4 static attractors.
}

