% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{krum_istate}
\alias{krum_istate}
\title{Initial state from Krumsiek et al.}
\format{A data frame with 1 row and 11 columns. 

Rows: each row consists of 1 set of Boolean state.
Columns: each column is for 1 gene/variable.}
\usage{
data(krum_istate)
}
\description{
An intial state specified in Krumsiek et al. It contains a set of Boolean values for 11 genes.
}

