% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculation.R
\name{m_score}
\alias{m_score}
\title{Inner function of calculating Boolean model score}
\usage{
m_score(x, bmodel, max_varperrule, detail = F)
}
\arguments{
\item{x}{matrix vector. Pairwise scores computed by dist_measure().}

\item{bmodel}{S4 BoolModel object. Model to be evaluated.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}

\item{detail}{logical. Whether to give more details in score calculation. Default to FALSE.}
}
\description{
This function calculates a final model score from pairwise and penalty scores.
}

