% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculation.R
\name{man_dist}
\alias{man_dist}
\title{Calculates pairwise Manhattan distances between two matrices}
\usage{
man_dist(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\description{
This function calculates pairwise Manhattan distances between two matrices.
}

