% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{match_state}
\alias{match_state}
\title{Check for matching states}
\usage{
match_state(mstate, xstate)
}
\arguments{
\item{mstate}{data frame. It should be a state(row) x gene(column) df. colnames will be used in comparison.}

\item{xstate}{data frame. It should be a state(row) x gene(column) df. colnames will be used in comparison.}
}
\description{
This function finds a match between two df of states. Returns a row index vector indicating for each row of mstate, what is the corresponding row in xstate. If a match cannot be found, a 0 will be return.
Only columns that are present in both df will be used in comparison. Note that the row index starts from 1 (as in R), not from 0 (as in cpp).
}

