% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_state_loop}
\alias{match_state_loop}
\title{Find a match between two data frames.}
\usage{
match_state_loop(mstate, xstate)
}
\arguments{
\item{mstate}{data frame. It should be a state(row) x gene(column) df.}

\item{xstate}{data frame. It should be a state(row) x gene(column) df.}
}
\description{
(&&&Not for public use&&&)This function finds a match between two df of states. Used in match_state(). Return an row index vector indicating which row of mstate matches the rows in xstate.
}

