% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_modification.R
\name{minmod_internal}
\alias{minmod_internal}
\title{Inner function of minimal modification of whole Boolean model}
\usage{
minmod_internal(bm, index)
}
\arguments{
\item{bm}{S4 BoolModel object.}

\item{index}{integer. Specifying rule of which gene to modify.}
}
\description{
This function generates all possible boolean models minimally modified. Returns a lists of 2 lists, deleted models and added models
}

