% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modification.R
\name{minmod_model}
\alias{minmod_model}
\title{Minimal modification of whole Boolean model}
\usage{
minmod_model(bm, index = NULL, overlap_gene = NULL)
}
\arguments{
\item{bm}{S4 BoolModel object.}

\item{index}{integer. Specifying rule of which gene to modify. If NULL, modifies all rules in the model. Defaults to NULL.}

\item{overlap_gene}{character vector. Specify which genes are present in both model and data inputs.}
}
\description{
This function generates all possible boolean models minimally modified. Returns a lists of 2 lists, deleted models and added models
}

