% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{model_setdiff}
\alias{model_setdiff}
\title{Find the set difference between two Boolean models}
\usage{
model_setdiff(x, y, max_varperrule, directed = F)
}
\arguments{
\item{x}{S4 BoolModel object. Test model.}

\item{y}{S4 BoolModel object. Reference model.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}

\item{directed}{logical. If TRUE, return the difference in terms with respect to x.}
}
\description{
This method takes in two models and find the set difference between them. Return a vector with the set difference.
}

