% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{model_simplify}
\alias{model_simplify}
\title{Simplifying Model}
\usage{
model_simplify(bmodel, istate, inter_bool, max_varperrule, verbose = F)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{istate}{data frame. Must have only 1 row, which represents 1 initial state.}

\item{inter_bool}{logical. Indicate whether to consider AND terms.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}

\item{verbose}{logical. Specifies whether to give detailed output. Default to F.}
}
\description{
This method takes in a model and remove redundant terms wrt to a single initial state.
Note that this model simplification is random, and the simplified model is not guaranteed to be the simplest model possible. It is only guaranteed to be a simpler model that can give the same state space as the orignal input model.
}

