% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{model_train}
\alias{model_train}
\title{Training Model}
\usage{
model_train(bmodel = NULL, edata, istate, max_varperrule = 6, tol = 1e-06,
  inter_bool, verbose = F, self_loop = F)
}
\arguments{
\item{bmodel}{Boolean model in data frame. If NULL, use a random Boolean model. Default to NULL.}

\item{edata}{list of 2 data frames. Initialised continuous and discretised expression data. Each data frame should have state(row) x gene(column).}

\item{istate}{data frame. Must have only 1 row, which represents 1 initial state.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must not be smaller than number of variables. Default to 6.}

\item{tol}{numeric. Specify the tolerance in ending condition. Default to 1e-6. It cannot be lower than .Machine$double.eps ^ 0.5.}

\item{inter_bool}{logical. Indicate whether to consider AND terms. Default to TRUE.}

\item{verbose}{logical. Specifies whether to give detailed output. Default to F.}

\item{self_loop}{logical. Indicates whether to allow self_loop in random starting model. Only used if is.null(bmodel). Default to F.}
}
\description{
This function performs model training to find the best model, using information from data. It requires an initial state supplied to perform the search, and an initial model can also be supplied to be included in the initial population.
Note that if a model is supplied, and the genes in the model is different from the genes in the data, only the genes overlapping between model and data will be retained for further analysis.
}

