% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.R
\name{model_train_sa}
\alias{model_train_sa}
\title{Training Model (using simulated annealing)}
\usage{
model_train_sa(edata, bmodel = NULL, istate = NULL, max_varperrule = 6,
  and_bool = T, self_loop = F, restart = 0, verbose = F)
}
\arguments{
\item{edata}{data frame of expression data. Should have state(row) x gene(column).}

\item{bmodel}{Boolean model in data frame. If NULL, use a random Boolean model. Defaults to NULL.}

\item{istate}{data frame. Must have only 1 row, which represents 1 initial state. Defaults to NULL.}

\item{max_varperrule}{integer. Maximum number of terms per rule (combining both act and inh rule). Note that this number must be higher than number of genes. Defaults to 6.}

\item{and_bool}{logical. Whether to consider AND terms. IF bmodel is not NULL, defaults to whether AND interaction is included in bmodel. If bmodel is NULL, then defaults to TRUE.}

\item{self_loop}{logical. Whether to allow self_loop in random starting model. Only used if is.null(bmodel). Default to F.}

\item{restart}{integer. Number of restart from the best solution. Defaults to 0.}

\item{verbose}{logical. Whether to give detailed output to the screen. Defaults to F.}
}
\description{
This function performs model training to find the best model, using information from data. It requires an initial state supplied to perform the search, and an initial model can also be supplied to be included in the initial population.
Note that if a model is supplied, and the genes in the model is different from the genes in the data, only the genes overlapping between model and data will be retained for further analysis.
}

