% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{moig_raw_data}
\alias{moig_raw_data}
\title{Raw single cell qRT-PCR expression data from Moignard et al.}
\format{A data frame with 597 rows and 18 columns.

Rows: each row consists of raw expression values from 1 cell.
Columns: each column is for 1 gene/variable.}
\usage{
data(moig_raw_data)
}
\description{
A raw single cell expression data obtained from multiple cell types.
}

