% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output_format.R
\name{outgenysis_model}
\alias{outgenysis_model}
\title{Output a Boolean Model into Genysis readable format}
\usage{
outgenysis_model(bmodel, filepath = getwd())
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{filepath}{character vector. Specify path (AND NOT file name). Default to current working directory, i.e. getwd(). Set to NULL to disable file output.}
}
\description{
This function outputs a Boolean Model in a format that is readable by Genysis. Return invisibly the formatted vector.
}

