% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_format.R
\name{outgraph_model}
\alias{outgraph_model}
\title{Output a Boolean Model into Cytoscape & Gephi readable format}
\usage{
outgraph_model(bmodel, path = getwd(), file = NULL, and_node = T)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{path}{character. Specify path (AND NOT file name). Default to current working directory, i.e. getwd(). Set to NULL to disable file output.}

\item{file}{character. Specify file name. Default to NULL for default file names.}

\item{and_node}{logical. Specify AND as an individual node. Default to T.}
}
\description{
This function outputs a Boolean Model in a format that is readable by Cytoscape and Gephi. Return invisibly the edges (with edge attributes) and node attributes. (i.e. list of 2 dfs)
}

