% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/general.R
\name{param_bimodal}
\alias{param_bimodal}
\title{Obtain parameters for bimodal distribution from real data}
\usage{
param_bimodal(x, data_type = "qpcr")
}
\arguments{
\item{x}{matrix. Input expression data. Col-genes, row-samples.}

\item{data_type}{character. Specify data types: qpcr, rnaseq.}
}
\description{
This function obtains parameters for bimodal distribution. Returns 4 parameters: mu1, mu2, sig1, sig2.
}

