% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotBM}
\alias{plotBM}
\title{Plot Boolean Model}
\usage{
plotBM(bmodel, makePlot = T, ...)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{makePlot}{logical. Whether to make plot or just return the object. Default to T.}

\item{...}{Additional parameters to plot.igraph.}
}
\description{
This method plots the network underlying Boolean models by using igraph for quick visualisation.
Require igraph.
}

