% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{printBM}
\alias{printBM}
\title{Print Boolean Model}
\usage{
printBM(bmodel, gene.names = F)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{gene.names}{logical. Specify whether to write rules in terms of genes or internal variables. Default to FALSE.}
}
\description{
This method converts the S4 BoolModel object back into a human-readable data frame, with two columns: (1) target genes, (2) Boolean rules.
}

