% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_m_score}
\alias{rcpp_m_score}
\title{Inner core for m_score()}
\usage{
rcpp_m_score(x_df)
}
\arguments{
\item{x_df}{matrix. Matrix with columns ranked wrt each row.}
}
\description{
This function takes in a df with columns ranked wrt each row, and try to assign each row to a unique column without repetition.
}

