% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_man_dist}
\alias{rcpp_man_dist}
\title{Calculating pairwise scores between model and data states.}
\usage{
rcpp_man_dist(x_df, y_df)
}
\arguments{
\item{x_df}{matrix. It should be numerical matrix of model states.}

\item{y_df}{matrix. It should be numerical matrix of data states.}
}
\description{
This function calculates the pairwise scores between each row of model and data states. The score is calculated using a custom binary distance measure.
}

