% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate}
\alias{rcpp_simulate}
\title{Simulate a Boolean model.}
\usage{
rcpp_simulate(bmodel, fstate, verbose = FALSE)
}
\arguments{
\item{bmodel}{list. A list of 4 lists created by decreate_model().}

\item{fstate}{data frame. It must have been initialised by initialise_data(), and has gene names as column names. Must contain only 1 row.}

\item{verbose}{logical. Indicates whether to output progress.}
}
\description{
(&&&Not for public use&&&)This function simulates the Boolean model using an initial state. For use within simulate_model(). Returns a matrix of full asynchronous state space.
}

