% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_validate}
\alias{rcpp_validate}
\title{Calculating validation scores between two adjacency matrices}
\usage{
rcpp_validate(inf_mat, true_mat)
}
\arguments{
\item{inf_mat}{matrix. It should be adjacency matrix of inferred network.}

\item{true_mat}{matrix. It should be adjacency matrix of true network.}
}
\description{
This function calculates the validation scores between two adjacency matrices.
}

