% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{real_param}
\alias{real_param}
\title{Estimated parameters from Wilson et al. data}
\format{A list with 4 numeric vectors, all_mu1, all_mu2, all_sig1, all_sig2. Note that each element in the vector is estimated from a single gene.}
\usage{
data(real_param)
}
\description{
A list of parameters (based on log normal distribution) estimated from Wilson et al. single-cell qPCR expression data.
}

