% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_model}
\alias{simulate_model}
\title{Simulating Boolean model}
\usage{
simulate_model(bmodel, istate, steady_bool = F)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{istate}{data frame. It must have been initialised by initialise_data(), and has gene names as column names. Must contain only 1 row.}

\item{steady_bool}{logical. Specifies whether to return point steady states or not. Default to F.}
}
\description{
This function simulates the Boolean model using an initial state. Returns the full asynchronous state space, and point steady states.
}

