% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{unique_raw_data}
\alias{unique_raw_data}
\title{Remove raw data duplicated wrt to the model state}
\usage{
unique_raw_data(dx, cx)
}
\arguments{
\item{dx}{matrix. Initialised and discretised numeric data of gene expression.}

\item{cx}{matrix. Initialised and continuous numeric data of gene expression.}
}
\description{
This function removes the 'duplicates' in an expression wrt to the model state.
}

