% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculation.R
\name{validate_adjmat}
\alias{validate_adjmat}
\title{Calculate true positive, true negative, false positive and false negative}
\usage{
validate_adjmat(inf_mat, true_mat)
}
\arguments{
\item{inf_mat}{matrix. It should be adjacency matrix of inferred network.}

\item{true_mat}{matrix. It should be adjacency matrix of true network.}
}
\description{
This function calculates the true positive, true negative, false positive and false negative values from the adjacency matrices.
}

