% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{vcat}
\alias{vcat}
\title{Verbose cat}
\usage{
vcat(string, bool)
}
\arguments{
\item{string}{character vector. String intended to be printed.}

\item{bool}{logical. Specify whether to print the string to stout or not.+}
}
\description{
This function is a simple wrapper for cat with a Boolean value for turning it on/off.
}

