% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{which.random.min}
\alias{which.random.min}
\title{Pick a random minimum value}
\usage{
which.random.min(x, favour_first = F)
}
\arguments{
\item{x}{numeric vector.}

\item{favour_first}{logical. If this is TRUE, and the first value in the vector supplies is one of the min, the index for the first value will always be returned.}
}
\description{
This function locates the minimum value in a vector (similar to which.min), however it will randomly break ties when there are multiple minimum values.
}

