% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{wilson_raw_data}
\alias{wilson_raw_data}
\title{Raw single cell qRT-PCR expression data from Wilson et al.}
\format{A data frame with 1626 rows and 44 columns. 

Rows: each row consists of raw expression values from 1 cell.
Columns: each column is for 1 gene/variable.}
\usage{
data(wilson_raw_data)
}
\description{
A raw single cell expression data obtained from multiple cell types.
}

