% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.R
\docType{data}
\name{wilson_raw_rnaseq}
\alias{wilson_raw_rnaseq}
\title{Raw single cell RNAseq expression data from Wilson et al.}
\format{A data frame with 96 rows and 38498 columns. 

Rows: each row consists of raw expression values from 1 cell.
Columns: each column is for 1 gene/variable.}
\usage{
data(wilson_raw_rnaseq)
}
\description{
A raw single cell expression data obtained from multiple cell types.
}

