% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{writeBM}
\alias{writeBM}
\title{Write Boolean Model}
\usage{
writeBM(bmodel, file, gene.names = F, rownames = F)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{file}{file name with path, or a file connection object.}

\item{gene.names}{logical. Specify whether to write rules in terms of genes or internal variables. Default to FALSE.}

\item{rownames}{logical. It specifies whether to write row names.}
}
\description{
This method writes the S4 BoolModel object into a CSV file. This method is a wrapper for print.BoolModel. The output is a data frame, with two columns: (1) target genes, (2) Boolean rules.
}

