% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_and}
\alias{check_and}
\title{Check if containing AND terms}
\usage{
check_and(bmodel)
}
\arguments{
\item{bmodel}{BoolModel object.}
}
\description{
This function checks if a particular Boolean model contains AND terms.
}

