% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{get_encodings}
\alias{get_encodings}
\title{Get corresponding encodings for compression or decompression.}
\usage{
get_encodings(bmodel, force_and = T)
}
\arguments{
\item{bmodel}{S4 BoolModel object.}

\item{force_and}{logical. Whether to include ANDs in the encoding even if no AND is present in the bmodel object. Defaults to T.}
}
\description{
This function gets all possible terms (single or double terms) and their corresponding encodings.
This function limits the number of possible variables in the model to 999.
}

