% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{initialise_raw_data}
\alias{initialise_raw_data}
\title{Initialise raw data}
\usage{
initialise_raw_data(x, max_expr = "high", uni_thre = 0.2, scale = T,
  discretised = F)
}
\arguments{
\item{x}{matrix. Numeric data of gene expression.}

\item{max_expr}{character. Specify whether max expression value is the lowest (as in qPCR), or the highest (as in RNAseq and microarray). Option: 'low', 'high'. Default to 'high'.}

\item{uni_thre}{numerical. Speficy threshold for unimodality test. Default to 0.2.}

\item{scale}{logical. Whether to scale the data to a range of 0-1. Default to T.}

\item{discretised}{logical. Whether to return discretised data. Default to F.}
}
\description{
This function initialise raw gene expression values in a matrix. Return either a matrix of (1) continuous values or (2) binary values.
Note that kmeans clustering as binarisation only works well if the data has a bimodal distribution.
}

