% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_modification.R
\name{minmod_internal}
\alias{minmod_internal}
\title{Inner function of minimal modification of whole Boolean model}
\usage{
minmod_internal(bm, index, encoded, model_encoding, and_bool, self_loop = F)
}
\arguments{
\item{bm}{S4 BoolModel object.}

\item{index}{integer. Specifying rule of which gene to modify.}

\item{encoded}{logical. Return Boolean models in encoded form to save space.}

\item{model_encoding}{list. Only used if encoded=T.}

\item{and_bool}{logical. Default to check_and() if unspecified.}

\item{self_loop}{logical. Whether to allow self_loop in random starting model. Default to F.}
}
\description{
This function generates all possible boolean models minimally modified. Returns a lists of 2 lists, deleted models and added models
}

