% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{model_consensus}
\alias{model_consensus}
\title{Intersection of genes}
\usage{
model_consensus(bmodel_list, model_encoding, format = "vec")
}
\arguments{
\item{bmodel_list}{list of BoolModel.}

\item{model_encoding}{list. Use for compressing and decompressing Boolean models.}

\item{format}{character. Specifies which format to return. Possible values: 'vec', 'df'. Default to 'vec'.}
}
\description{
This function finds the intersection of genes and provide a score for them. Return a consensus model or a vector of scores.
}

