% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{model_dist}
\alias{model_dist}
\title{Calculate distance between Boolean models}
\usage{
model_dist(x, y)
}
\arguments{
\item{x}{S4 BoolModel object. Test model.}

\item{y}{S4 BoolModel object. Reference model.}
}
\description{
This method takes in two models and calculate the distance between them. The value return indicate the number of steps between the two models.
}

